+++ tests/rdt/transformers/test_NumberTransformer.py
@@ -6,7 +6,8 @@ import pandas as pd
 from rdt.transformers.NumberTransformer import NumberTransformer
 
 
-class Test_NumberTransformer(unittest.TestCase):
+class TestNumberTransformer(unittest.TestCase):
+
     def test___init__(self):
         """On init, sets type to number"""
 
@@ -14,76 +15,102 @@ class Test_NumberTransformer(unittest.TestCase):
         transformer = NumberTransformer()
 
         # Check
-        assert (transformer.type == 'number')
+        assert transformer.type == 'number'
 
     def test_fit_transform(self):
-        """FIXME: finish when number and null transformer are compatible"""
-        """ fit_transform sets internal state and transforms data """
-
+        """fit_transform sets internal state and transforms data."""
         # Setup
-        col = pd.Series([62, 27, 5, 34, 62])
+        col = pd.DataFrame({
+            'age': [62, 27, 5, 34, 62]
+        })
+
         col_meta = {
             "name": "age",
             "type": "number",
             "subtype": "integer",
         }
         transformer = NumberTransformer()
-        data = {'age': [62, 27, 5, 34, 62]}
+        expected_result = pd.DataFrame({'age': [62, 27, 5, 34, 62]})
 
         # Run
-        result = pd.DataFrame(data)
-        expected_result = transformer.fit_transform(col, col_meta, False)
+        result = transformer.fit_transform(col, col_meta, False)
 
         # Check
         assert result.equals(expected_result)
 
-    @unittest.skip("FIXME: when number and null transformer are compatible")
     def test_fit_transform_missing(self):
-        """FIXME: finish when number and null transformer are compatible"""
-        """ Sets internal state and transforms data with missing values"""
-
+        """Sets internal state and transforms data with missing values."""
         # Setup
-        col = pd.Series([62, 27, np.nan, 34, 62])
+        col = pd.DataFrame({
+            'age': [62, 27, np.nan, 34, 62],
+        })
         col_meta = {
             "name": "age",
             "type": "number",
             "subtype": "integer",
         }
         transformer = NumberTransformer()
-        data = {'age': [62, 27, 0, 34, 62]}
+        expected_result = pd.DataFrame({
+            'age': [62, 27, 46, 34, 62],
+            '?age': [1, 1, 0, 1, 1]
+        })
 
         # Run
-        result = pd.DataFrame(data)
-        expected_result = transformer.fit_transform(col, col_meta, True)
+        result = transformer.fit_transform(col, col_meta, True)
 
         # Check
-        assert result.age.equals(expected_result.age)
+        assert result.equals(expected_result)
 
     def test_reverse_transform(self):
-        """FIXME: finish when number and null transformer are compatible"""
-        """ Checks the conversion of the data back into original format """
-
+        """Checks the conversion of the data back into original format."""
         # Setup
-        col = pd.Series([34, 23, 27, 31, 39], name='age')
+        col = pd.DataFrame({
+            'age': [34, 23, 27, 31, 39]
+        })
         col_meta = {
             'name': 'age',
             'subtype': 'integer',
             'type': 'number'
         }
         transformer = NumberTransformer()
+        expected_result = pd.DataFrame({'age': [34, 23, 27, 31, 39]})
 
         # Run
-        result = pd.DataFrame({'age': [34, 23, 27, 31, 39]})
-        expected_result = transformer.reverse_transform(col, col_meta, False)
+        result = transformer.reverse_transform(col, col_meta, False)
 
         # Check
         assert result.equals(expected_result)
 
-    @unittest.skip("FIXME: when number and null transformer are compatible")
-    def test_reverse_transform_missing(self):
-        """FIXME: finish when number and null transformer are compatible"""
-        """ Sets internal state and transforms data with missing values"""
+    def test_reverse_transform_nan(self):
+        """Checks that nans are handled correctly in reverse transformation."""
+        # Setup
+        col = pd.DataFrame({
+            'age': [34, 23, 27, 31, 39]
+        })
+        col_meta = {
+            'name': 'age',
+            'subtype': 'integer',
+            'type': 'number'
+        }
+        transformer = NumberTransformer()
+        transformer.fit_transform(col, col_meta, False)
 
+        col2 = pd.DataFrame({
+            'age': [0, 10, 20, 30, np.nan]
+        })
+
+        expected_result = pd.DataFrame({
+            'age': [0, 10, 20, 30, transformer.default_val]
+        })
+
+        # Run
+        result = transformer.reverse_transform(col2, col_meta, False)
+
+        # Check
+        assert result.equals(expected_result)
+
+    def test_reverse_transform_missing(self):
+        """Sets internal state and transforms data with missing values."""
         # Setup
         col = pd.DataFrame({
             'age': [34, 23, 0, 31, 39],
@@ -95,16 +122,16 @@ class Test_NumberTransformer(unittest.TestCase):
             'type': 'number'
         }
         transformer = NumberTransformer()
+        expected_result = pd.DataFrame({'age': [34, 23, np.nan, 31, 39]})
 
         # Run
-        result = pd.DataFrame({'age': [34, 23, np.nan, 31, 39]})
-        expected_result = transformer.reverse_transform(col, col_meta, True)
+        result = transformer.reverse_transform(col, col_meta, True)
 
         # Check
         assert result.equals(expected_result)
 
     def test_get_val_subtype_integer(self):
-        """ Rounds off the value and returns it """
+        """Rounds off the value and returns it."""
 
         # Setup
         col = pd.DataFrame({
@@ -124,8 +151,7 @@ class Test_NumberTransformer(unittest.TestCase):
         assert result.equals(expected_result)
 
     def test_get_val_subtype_not_integer(self):
-        """ If subtype is not integer and there are not null values,
-        returns the same value """
+        """If subtype is not integer and there are not null values, returns the same value."""
 
         # Setup
         col = pd.DataFrame({
@@ -145,8 +171,7 @@ class Test_NumberTransformer(unittest.TestCase):
         assert result.equals(expected_result)
 
     def test_get_val_null_value(self):
-        """ If subtype is not integer and there are null values,
-         returns the default value """
+        """get_val return the default value for null values if subtype is not integer."""
 
         # Setup
         col = pd.DataFrame({
@@ -156,8 +181,8 @@ class Test_NumberTransformer(unittest.TestCase):
         transformer = NumberTransformer()
         transformer.col_name = 'age'
         transformer.subtype = 'decimal'
-        transformer.default_val = col.age[0]
-        expected_result = pd.Series([4.0, 4.0, 4.0])
+        transformer.default_val = 5.0
+        expected_result = pd.Series([4.0, 5.0, 5.0])
         # Run
         result = col.apply(transformer.get_val, axis=1)
 
@@ -165,7 +190,7 @@ class Test_NumberTransformer(unittest.TestCase):
         assert result.equals(expected_result)
 
     def test_get_val_error(self):
-        """ Returns 'default_val' when there is a ValueError """
+        """Returns 'default_val' when there is a ValueError."""
         # Setup
         col = pd.DataFrame({
             'age': [4, 'hoo', 13]
@@ -184,7 +209,7 @@ class Test_NumberTransformer(unittest.TestCase):
         assert result.equals(expected_result)
 
     def test_get_number_converter(self):
-        """ If meta 'integer', cast values to intif not, returns as is """
+        """If meta 'integer', cast values to int."""
         # Setup
         transformer = NumberTransformer()
         col_name = 'age'