+++ HISTORY.md
@@ -1,5 +1,16 @@
 # History
 
+## 0.1.1
+
+* Improve handling of NaN in NumberTransformer and CatTransformer.
+* Add unittests for HyperTransformer.
+* Remove unused methods `get_types` and `impute_table` from HyperTransformer.
+* Make NumberTransformer enforce dtype int on integer data.
+* Make DTTransformer check data format before transforming.
+* Add minimal API Reference.
+* Merge `rdt.utils` into `HyperTransformer` class. 
+
+
 ## 0.1.0
 
 * First release on PyPI.